COMMERCE-6529 move
